+++ jest.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  coverageDirectory: 'coverage',
+  coverageProvider: 'v8',
+  coverageReporters: ['json', 'text'],
+  testTimeout: 30000,
+}